db_core/requirements.txt
aioredis>=1.3.1
asyncio-nats-client>=0.10.0
pymongo>=3.10.1
